+++ tests/test_anonymize.py
@@ -1,10 +1,6 @@
 import os
 import sys
 import pytest
-try:
-    from StringIO import StringIO
-except:
-    from io import StringIO
 from anonymize import Anonymize
 
 
@@ -12,15 +8,6 @@ from anonymize import Anonymize
 def sample():
     BASE_DIR = os.path.dirname(os.path.dirname(__file__))
     return os.path.join(BASE_DIR, 'anonymize', "sample1.yml")
-
-
-@pytest.fixture
-def stdout():
-    old_stdout = sys.stdout
-
-    result = StringIO()
-    sys.stdout = result
-    return old_stdout, result
 
 
 def test_should_get_trucate_at_output(sample, stdout):
